lift weatherData and selectedLocation up
