+++ src/App.interfaces.ts
@@ -1,0 +1,7 @@
+export interface WeatherData {
+  temperature_2m_max: number[];
+  temperature_2m_min: number[];
+  time: number[];
+  wind_speed_10m_max: number[];
+  weathercode: number[];
+}
